pjt/urls.py
from django.contrib import admin
from django.urls import path, include
# 이미지 업로드
from django.conf import settings
from django.conf.urls.static import static


urlpatterns = [
    path('admin/', admin.site.urls),
    path('posts/', include('posts.urls')),
    path('accounts/', include('accounts.urls')),
] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)

if settings.DEBUG:
    import debug_toolbar
    urlpatterns += [
            path('__debug__/', include(debug_toolbar.urls)),
        ]